extract hook; fix autosave
