+++ core/app/c/(public)/[communitySlug]/public/forms/[formSlug]/fill/ExternalFormWrapper.tsx
@@ -20,6 +20,7 @@ import type { GetPubResponseBody, JsonValue } from "contracts";
 import type { PubsId, PubTypesId } from "db/public";
 import { CoreSchemaType, ElementType } from "db/public";
 import { Form } from "ui/form";
+import { useUnsavedChangesWarning } from "ui/hooks";
 import { cn } from "utils";
 
 import type { FormElementToggleContext } from "~/app/components/forms/FormElementToggleContext";
@@ -35,20 +36,6 @@ import { SAVE_STATUS_QUERY_PARAM, SUBMIT_ID_QUERY_PARAM } from "./constants";
 import { SubmitButtons } from "./SubmitButtons";
 
 const SAVE_WAIT_MS = 5000;
-
-const useUnsavedChangesWarning = ({ isDirty }: FormState<FieldValues>) => {
-	useEffect(() => {
-		const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-			if (isDirty) {
-				event.preventDefault();
-			}
-		};
-		window.addEventListener("beforeunload", handleBeforeUnload);
-		return () => {
-			window.removeEventListener("beforeunload", handleBeforeUnload);
-		};
-	}, [isDirty]);
-};
 
 const isComplete = (formElements: PubPubForm["elements"], values: FieldValues) => {
 	const requiredElements = formElements.filter((fe) => fe.required && fe.slug);
@@ -82,7 +69,12 @@ const preparePayload = ({
 	// we do not want to pass an empty `email` field to the upsert (it will fail validation)
 	const payload: Record<string, JsonValue> = {};
 	for (const { slug } of formElements) {
-		if (slug && toggleContext.isEnabled(slug) && formState.dirtyFields[slug]) {
+		if (
+			slug &&
+			toggleContext.isEnabled(slug) &&
+			// Only send fields that were changed.
+			formState.dirtyFields[slug]
+		) {
 			payload[slug] = formValues[slug];
 		}
 	}
@@ -240,6 +232,19 @@ export const ExternalFormWrapper = ({
 				if (!isUpdating) {
 					newParams.set("pubId", pubId);
 				}
+
+				if (autoSave) {
+					// Reset dirty state to prevent the unsaved changes warning from
+					// blocking navigation.
+					// See https://stackoverflow.com/questions/63953501/react-hook-form-resetting-isdirty-without-clearing-form
+					formInstance.reset(
+						{},
+						{
+							keepValues: true,
+						}
+					);
+				}
+
 				if (!autoSave && isComplete(formElements, pubValues)) {
 					if (submitButtonId) {
 						newParams.set(SUBMIT_ID_QUERY_PARAM, submitButtonId);
@@ -268,7 +273,7 @@ export const ExternalFormWrapper = ({
 		formInstance.trigger(Object.keys(formInstance.formState.errors));
 	}, [formInstance, toggleContext]);
 
-	useUnsavedChangesWarning(formInstance.formState);
+	useUnsavedChangesWarning(formInstance.formState.isDirty);
 
 	const isSubmitting = formInstance.formState.isSubmitting;
